+++ pkg/util/kconfig/client.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2025 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,7 +18,7 @@
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
 
-package kclient
+package kconfig
 
 import (
 	"fmt"
@@ -32,8 +32,8 @@ import (
 
 const Kubeconfig util.EnvironmentVariable = "KUBECONFIG"
 
-// newKubeConfig loads config from KUBECONFIG or as incluster
-func newKubeConfig() (*rest.Config, error) {
+// NewConfig loads config from KUBECONFIG or as incluster
+func NewConfig() (*rest.Config, error) {
 	// If KUBECONFIG is defined use this variable
 	if kubeconfig, ok := Kubeconfig.Lookup(); ok {
 		return clientcmd.BuildConfigFromFlags("", kubeconfig)